Set hover effect
I set up a hover effect so that the grid divs change color when my mouse passes over them, leaving a
pixelated trail through the grid. I also added a button to the top of the screen that sends the user a
asking for the number of sides for the new grid. Once entered, the existing grid will be removed and a new
grid will be generated in the same total space as before.